Refactor x402-faremeter example to separate client from demo
Extract protocol internals (gasless handler, v2 normalization) into
x402-client.ts, exposing a single createX402Client() factory. This
makes the agent integration pattern clearer in index.ts.
